requirements.txt
tensorflow==2.17.0
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
opencv-python
ipywidgets
streamlit
